# backend/tests/api/fixtures/openapi_v1.snapshot.json
{
  "components": {
    "schemas": {
...
      "ActivityCountResponse": {
        "description": "Count of activities response schema.",
        "properties": {
          "count": {
            "description": "Total number of activity records",
            "examples": [
              42
            ],
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "Activity.CountResponse",
        "type": "object"
      },
      "ActivityListResponse": {
        "description": "List of activities for GET responses.",
        "properties": {
          "activities": {
            "description": "List of activities",
            "items": {
              "$ref": "#/components/schemas/ActivityResponse"
            },
            "title": "Activities",
            "type": "array"
          }
        },
        "required": [
          "activities"
        ],
        "title": "Activity.ListResponse",
        "type": "object"
      },
...
      "ActivityResponse": {
        "description": "Activity response schema.",
        "properties": {
          "activityId": {
            "description": "Functional ID identifying this activity",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550E8400-E29B-41D4-A716-446655440000"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Activityid",
            "type": "string"
          },
          "activityName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display name (optional) of the activity",
            "title": "Activityname"
          },
          "address": {
            "$ref": "#/components/schemas/CommonAddressResponse",
            "description": "Address composite (`thoroughfare`, `locatorDesignatorNumber` (optional), `locatorDesignatorLetter` (optional), `locatorDesignatorAddition` (optional), `postCode`, `postName`, `fullAddress`)"
          },
          "areaId": {
            "description": "Functional ID referencing the area where this activity took place",
            "examples": [
              "3ab7c2b9-5c8d-4100-bc3e-00ac115f0495",
              "3AB7C2B9-5C8D-4100-BC3E-00AC115F0495"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Areaid",
            "type": "string"
          },
          "competentAuthorityId": {
            "description": "Functional ID referencing the competent authority that owns the area",
            "examples": [
              "sdep-ca0363",
              "SDEP-CA0363"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Competentauthorityid",
            "type": "string"
          },
          "competentAuthorityName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display name (optional) of the competent authority",
            "title": "Competentauthorityname"
          },
          "countryOfGuests": {
            "description": "Array of country codes of guests (each element is ISO 3166-1 alpha-3 or 'N/A'); array length equals numberOfGuests.",
            "items": {
              "type": "string"
            },
            "title": "Countryofguests",
            "type": "array"
          },
          "createdAt": {
            "description": "Timestamp when this activity version was created (UTC)",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "numberOfGuests": {
            "description": "Number of guests (1-1024)",
            "title": "Numberofguests",
            "type": "integer"
          },
          "platformId": {
            "description": "Functional ID referencing the platform that owns the activity",
            "examples": [
              "str01",
              "STR01"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Platformid",
            "type": "string"
          },
          "platformName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display name (optional) of the platform",
            "title": "Platformname"
          },
          "registrationNumber": {
            "description": "Registration number of the address",
            "title": "Registrationnumber",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ActivityStatus",
            "description": "Lifecycle status of the activity: `finished` or `cancelled`.",
            "examples": [
              "finished",
              "cancelled"
            ]
          },
          "temporal": {
            "$ref": "#/components/schemas/CommonTemporalResponse",
            "description": "Temporal composite (`startDatetime`, `endDatetime`)"
          },
          "url": {
            "description": "URL of the originating listing/advertisement",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "activityId",
          "status",
          "areaId",
          "competentAuthorityId",
          "url",
          "address",
          "registrationNumber",
          "numberOfGuests",
          "countryOfGuests",
          "temporal",
          "platformId",
          "createdAt"
        ],
        "title": "Activity.Response",
        "type": "object"
      },
      "ActivityStatus": {
        "description": "Lifecycle status for an activity record.",
        "enum": [
          "finished",
          "cancelled"
        ],
        "title": "Activity.Status",
        "type": "string"
      },
...
      "CommonAddressResponse": {
        "description": "Address composite schema for activity responses (INSPIRE/STR-AP field names).",
        "properties": {
          "fullAddress": {
            "description": "Full address as a single string",
            "title": "Fulladdress",
            "type": "string"
          },
          "locatorDesignatorAddition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional qualifier (optional)",
            "title": "Locatordesignatoraddition"
          },
          "locatorDesignatorLetter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Letter/character suffix (optional)",
            "title": "Locatordesignatorletter"
          },
          "locatorDesignatorNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Numeric house number component (optional)",
            "title": "Locatordesignatornumber"
          },
          "postCode": {
            "description": "Postal code",
            "title": "Postcode",
            "type": "string"
          },
          "postName": {
            "description": "City / town / village",
            "title": "Postname",
            "type": "string"
          },
          "thoroughfare": {
            "description": "Street / public space name",
            "title": "Thoroughfare",
            "type": "string"
          }
        },
        "required": [
          "thoroughfare",
          "postCode",
          "postName",
          "fullAddress"
        ],
        "title": "Common.AddressResponse",
        "type": "object"
      },
...
      "CommonTemporalResponse": {
        "description": "Temporal composite schema for activity responses.",
        "properties": {
          "endDatetime": {
            "description": "End date and time of the rental activity",
            "format": "date-time",
            "title": "Enddatetime",
            "type": "string"
          },
          "startDatetime": {
            "description": "Start date and time of the rental activity",
            "format": "date-time",
            "title": "Startdatetime",
            "type": "string"
          }
        },
        "required": [
          "startDatetime",
          "endDatetime"
        ],
        "title": "Common.TemporalResponse",
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Detail of a single error.",
        "properties": {
          "loc": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Location of the error (e.g. ['body', 'fieldName'])",
            "title": "Loc"
          },
          "msg": {
            "description": "Human-readable error message",
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "description": "Error type identifier",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "msg",
          "type"
        ],
        "title": "Error.Detail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Standardized error response format.",
        "properties": {
          "detail": {
            "description": "List of error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error.Response",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2ClientCredentials": {
        "flows": {
          "clientCredentials": {
            "scopes": {},
            "tokenUrl": "<token-url>"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "SDEP is a gateway for the electronic transmission of data between online short-term rental platforms (STR) and competent authorities (CA), ensuring timely, reliable and efficient data sharing processes.\n\n- [Github (open source)](https://github.com/SEMICeu/sdep)\n\n- [EU legislation](https://eur-lex.europa.eu/eli/reg/2024/1028/oj/eng)\n\n- [STR Application Profile (STR-AP)](https://semiceu.github.io/STR-AP/releases/1.0.1/)\n\n- [STR origin prototype v0.0.67](https://eu-str.sdep-pilot.eu/swagger/index.html)\n\nContact:\n\n- [boris.dijkmans@rijksoverheid.nl](mailto:boris.dijkmans@rijksoverheid.nl)",
    "title": "Short Term Rental (STR) - Single Digital Entry Point (SDEP)",
    "version": "<frozen-version>"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ca/activities": {
      "get": {
        "description": "Get activities for the currently authenticated competent authority. By default, returns all current activities (unlimited), including current records whose lifecycle `status` is `cancelled`. Use optional pagination parameters to limit results.\n\n**Each activity contains:**\n- `activityId`: Functional ID identifying this activity\n- `activityName`: Display name (optional) of the activity\n- `status`: Lifecycle status of the activity: `finished` or `cancelled`\n- `areaId`: Functional ID referencing the area where the activity took place\n- `competentAuthorityId`: Functional ID referencing the competent authority that owns the area\n- `competentAuthorityName`: Display name (optional) of the competent authority\n- `url`: URL of the originating listing/advertisement\n- `address`: Address composite (`thoroughfare`, `locatorDesignatorNumber` (optional), `locatorDesignatorLetter` (optional), `locatorDesignatorAddition` (optional), `postCode`, `postName`, `fullAddress`)\n- `registrationNumber`: Registration number of the address\n- `numberOfGuests`: Number of guests (1-1024)\n- `countryOfGuests`: Array of country codes of guests (each element is ISO 3166-1 alpha-3 or `N/A`; array length equals `numberOfGuests`)\n- `temporal`: Temporal composite (`startDatetime`, `endDatetime`)\n- `platformId`: Functional ID referencing the platform that owns the activity\n- `platformName`: Display name (optional) of the platform\n- `createdAt`: Timestamp when this activity version was created (UTC)",
        "operationId": "getActivityByCompetentAuthority",
        "parameters": [
          {
            "description": "Number of records to skip (default: 0)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of records to skip (default: 0)",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of records to return (default: unlimited, max: 1000 when specified)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of records to return (default: unlimited, max: 1000 when specified)",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "activities": [
                    {
                      "activityId": "550e8400-e29b-41d4-a716-446655440000",
                      "activityName": "Amsterdam Summer Rental",
                      "address": {
                        "fullAddress": "Prinsengracht 263, 1016GV Amsterdam",
                        "locatorDesignatorNumber": 263,
                        "postCode": "1016GV",
                        "postName": "Amsterdam",
                        "thoroughfare": "Prinsengracht"
                      },
                      "areaId": "959a7439-7cad-4009-96ec-353b44723db9",
                      "competentAuthorityId": "sdep-ca0363",
                      "competentAuthorityName": "Gemeente Amsterdam",
                      "countryOfGuests": [
                        "NLD",
                        "DEU",
                        "BEL",
                        "N/A"
                      ],
                      "createdAt": "2025-06-01T12:00:00Z",
                      "numberOfGuests": 4,
                      "platformId": "sdep-str01",
                      "platformName": "Test STR 01 (interactive usage, persistent)",
                      "registrationNumber": "REG0001",
                      "status": "finished",
                      "temporal": {
                        "endDatetime": "2025-06-07T11:00:00Z",
                        "startDatetime": "2025-06-01T14:00:00Z"
                      },
                      "url": "http://example.com/amsterdam-myhouse-1"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ActivityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request - invalid query parameters"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get activities for the currently authenticated competent authority",
        "tags": [
          "ca"
        ]
      }
    },
    "/ca/activities/count": {
      "get": {
        "description": "Get activities count for the currently authenticated competent authority (optional, to support pagination). Counts all current activity records, including those whose lifecycle `status` is `cancelled`.",
        "operationId": "countActivities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityCountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get activities count for the currently authenticated competent authority (optional, to support pagination)",
        "tags": [
          "ca"
        ]
      }
    },
...
  }
}